+++ cpp/cluster/Cluster.cc
@@ -71,14 +71,15 @@ uint32_t DisjointSet::find(const uint32_t c)
     return r;
     }
 
-Cluster::Cluster(const box::Box& box, float rcut)
-    : m_box(box), m_rcut(rcut), m_num_particles(0)
+Cluster::Cluster(float rcut)
+    : m_rcut(rcut), m_num_particles(0)
     {
     if (m_rcut < 0.0f)
         throw invalid_argument("rcut must be positive");
     }
 
-void Cluster::computeClusters(const freud::locality::NeighborList *nlist,
+void Cluster::computeClusters(const box::Box& box,
+                              const freud::locality::NeighborList *nlist,
                               const vec3<float> *points,
                               unsigned int Np)
     {
@@ -112,12 +113,8 @@ void Cluster::computeClusters(const freud::locality::NeighborList *nlist,
                     {
                     // compute r between the two particles
                     vec3<float> delta = p - points[j];
-                    // float dx = float(p.x - points[j].x);
-                    // float dy = float(p.y - points[j].y);
-                    // float dz = float(p.z - points[j].z);
-                    delta = m_box.wrap(delta);
+                    delta = box.wrap(delta);
 
-                    // float rsq = delta.x*delta.x + delta.y*delta.y + delta.z*delta.z;
                     float rsq = dot(delta, delta);
                     if (rsq < rmaxsq)
                         {